Move Doi and Orcid out into model/mod.rs (no changes) for shareability across objects
