+++ thoth-app/src/component/work.rs
@@ -5,12 +5,12 @@ use thoth_api::funding::model::FundingExtended as Funding;
 use thoth_api::imprint::model::ImprintExtended as Imprint;
 use thoth_api::issue::model::IssueExtended as Issue;
 use thoth_api::language::model::Language;
+use thoth_api::model::{Doi, DOI_DOMAIN};
 use thoth_api::publication::model::PublicationExtended as Publication;
 use thoth_api::subject::model::Subject;
 use thoth_api::work::model::WorkExtended as Work;
 use thoth_api::work::model::WorkStatus;
 use thoth_api::work::model::WorkType;
-use thoth_api::work::model::{Doi, DOI_DOMAIN};
 use uuid::Uuid;
 use yew::html;
 use yew::prelude::*;
