+++ thoth-app/src/models/contributor/update_contributor_mutation.rs
@@ -1,6 +1,7 @@
 use serde::Deserialize;
 use serde::Serialize;
-use thoth_api::contributor::model::{Contributor, Orcid};
+use thoth_api::contributor::model::Contributor;
+use thoth_api::model::Orcid;
 use uuid::Uuid;
 
 const UPDATE_CONTRIBUTOR_MUTATION: &str = "
