+++ thoth-app/src/models/work/update_work_mutation.rs
@@ -1,6 +1,6 @@
 use serde::Deserialize;
 use serde::Serialize;
-use thoth_api::work::model::Doi;
+use thoth_api::model::Doi;
 use thoth_api::work::model::WorkStatus;
 use thoth_api::work::model::WorkType;
 use uuid::Uuid;
